+++ CncControlerGui/UpdateManagerThread.cpp
@@ -166,9 +166,9 @@ void UpdateManagerThread::popAndFormatPosSpyQueue() {
 								posSpyRow.updateItem(CncPosSpyListCtrl::COL_REF, 	wxString::Format(rFormat, 		lpse.pos.id));
 								posSpyRow.updateItem(CncPosSpyListCtrl::COL_T, 		wxString::Format("%c", 			lpse.pos.speedMode));
 								posSpyRow.updateItem(CncPosSpyListCtrl::COL_F, 		speedInfo);
-								posSpyRow.updateItem(CncPosSpyListCtrl::COL_X, 		wxString::Format("%ld", 		lpse.pos.pos.getX()));
-								posSpyRow.updateItem(CncPosSpyListCtrl::COL_Y, 		wxString::Format("%ld", 		lpse.pos.pos.getY()));
-								posSpyRow.updateItem(CncPosSpyListCtrl::COL_Z, 		wxString::Format("%ld", 		lpse.pos.pos.getZ()));
+								posSpyRow.updateItem(CncPosSpyListCtrl::COL_X, 		wxString::Format("%ld", 		(long)lpse.pos.pos.getX()));
+								posSpyRow.updateItem(CncPosSpyListCtrl::COL_Y, 		wxString::Format("%ld", 		(long)lpse.pos.pos.getY()));
+								posSpyRow.updateItem(CncPosSpyListCtrl::COL_Z, 		wxString::Format("%ld", 		(long)lpse.pos.pos.getZ()));
 		}
 		
 		posSpyStringQueue.push(posSpyRow);
@@ -201,9 +201,9 @@ void UpdateManagerThread::popAndFormatSetterQueue() {
 				case SEPARARTOR_RESET:		label.assign(wxString::Format("Bookmark: Type(<RESET>[%s])", portName)); break;
 				case SEPARARTOR_RUN:		label.assign(wxString::Format("Bookmark: Type(<RUN>[%s])",   portName)); break;
 			}
-			setterRow.updateItem(CncSetterListCtrl::COL_TYPE, 		wxString::Format("%ld", value));
+			setterRow.updateItem(CncSetterListCtrl::COL_TYPE, 		wxString::Format("%ld", 		(long)value));
 			setterRow.updateItem(CncSetterListCtrl::COL_KEY, 		label);
-			setterRow.updateItem(CncSetterListCtrl::COL_VAL, 		wxString::Format("%s.%03ld", 	lste.ts.FormatTime(), lste.ts.GetMillisecond()));
+			setterRow.updateItem(CncSetterListCtrl::COL_VAL, 		wxString::Format("%s.%03ld", 	lste.ts.FormatTime(), (long)lste.ts.GetMillisecond()));
 			
 		} else {
 			setterRow.updateItem(CncSetterListCtrl::COL_TYPE, 		"");
